pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "solaihack2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "hydra-colorlog==1.2.0",
    "hydra-core==1.3.2",
    "hydra-optuna-sweeper==1.2.0",
    "lightning>=2.0.0",
    "pandas>=2.2.3",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
    "rich>=13.9.4",
    "rootutils>=1.0.7",
    "torch>=2.0.0",
    "torchmetrics>=0.11.4",
    "torchvision>=0.15.0",
]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
]
